Move translateAllFramesInTab to separate file
